+++ src/main/java/org/dev/powermarket/service/dto/RentalRequestDto.java
@@ -30,5 +30,5 @@ public class RentalRequestDto {
     private String rejectionReason;
     private Instant createdAt;
     private Instant respondedAt;
-    private Integer capacityNeeded;
+    private double capacityNeeded;
 }
